+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+- `egg` now provides `Symbol`s, a simple interned string that users can (and
+  should) use in their `Language`s.
+- `egg` will now warn you when you try to use `Rewrite`s with the same name.
+  
+### Fixed
+- The `rewrite!` macro builds conditional rewrites in the correct order now.
+
 ## [0.4.1] - 2020-05-26
 
 ### Added